refactor: extract SettingsReadinessTypes.swift from SettingsReadiness.swift
Extracts 2 support types (SettingsReadinessStatus enum + PrerequisiteRow struct)
from SettingsReadiness.swift into a new SettingsReadinessTypes.swift.

Byte-identical move; SHA verified against origin/main.

Imports: SwiftUI (for Color used by SettingsReadinessStatus).

Reduces SettingsReadiness.swift from 218 to 160 lines. Only the primary
SettingsReadiness enum (the readiness computation logic) remains.

Tests: 667 (unchanged).
