+++ Sources/BugNarrator/Views/Settings/SettingsReadiness.swift
@@ -5,64 +5,6 @@ import SwiftUI
 /// top-level "at a glance" status summary and the Integrations pane can compute
 /// and render the same status rows. Pure relocation: identical logic and layout.
 
-/// The configured-ness of a Settings capability (AI provider, GitHub, Jira).
-enum SettingsReadinessStatus {
-    case ready
-    case needsSetup
-    case pendingSave
-    case locked
-
-    var title: String {
-        switch self {
-        case .ready:
-            return "Ready"
-        case .needsSetup:
-            return "Needs setup"
-        case .pendingSave:
-            return "Pending save"
-        case .locked:
-            return "Locked"
-        }
-    }
-
-    var symbolName: String {
-        switch self {
-        case .ready:
-            return "checkmark.circle.fill"
-        case .needsSetup:
-            return "exclamationmark.circle.fill"
-        case .pendingSave:
-            return "clock.fill"
-        case .locked:
-            return "lock.fill"
-        }
-    }
-
-    var color: Color {
-        switch self {
-        case .ready:
-            return .green
-        case .needsSetup:
-            return .orange
-        case .pendingSave:
-            return .blue
-        case .locked:
-            return .red
-        }
-    }
-}
-
-/// One row in a per-integration prerequisite checklist.
-struct PrerequisiteRow: Identifiable {
-    let title: String
-    let detail: String
-    let status: SettingsReadinessStatus
-
-    var id: String {
-        title
-    }
-}
-
 /// Pure readiness computations derived from `SettingsStore` state.
 enum SettingsReadiness {
     static func credentialStatus(
